+++ app/core/constants.py
@@ -1,0 +1,7 @@
+practice_area_lead = "PracticeAreaLead"
+global_admin = "GlobalAdmin"
+project_team_member = "ProjectTeamMember"
+practice_area_team_member = "PracticeAreaTeamMember"
+verified_user = "VerifiedUser"
+project_lead = "ProjectLead"
+
