+++ app/core/tests/security/constants.py
@@ -1,0 +1,23 @@
+
+practice_area_lead = "PracticeAreaLead"
+global_admin = "GlobalAdmin"
+project_team_member = "ProjectTeamMember"
+practice_area_team_member = "PracticeAreaTeamMember"
+verified_user = "VerifiedUser"
+project_lead = "ProjectLead"
+
+website_project = "website"
+people_depot_project = "people-depot"
+wanda_name = "Wanda"
+wally_name = "Wally"
+winona_name = "Winona"
+zani_name = "Zani"
+patti_name = "Patti"
+patrick_name = "Patrick"
+paul_name = "Paul"
+garry_name = "Garry"
+valerie_name = "Valerie"
+
+
+
+
