--- app/core/tests/test_populate.py
@@ -1,7 +1,0 @@
-from django.test import TestCase
-from .base_test_case import BaseTestCase
-
-class TestPopulate(BaseTestCase):
-    def test_populate(self):
-        print("TestPopulate: test_populate")
-        pass
