--- app/data/group_data.py
@@ -1,7 +1,0 @@
-PROJECT_LEAD = "ProjectLead"
-PRACTICE_AREA_LEAD = "PracticeAreaLead"
-GLOBAL_ADMIN = "GlobalAdmin"
-PROJECT_TEAM_MEMBER = "ProjectTeamMember"
-PRACTICE_AREA_TEAM_MEMBER = "PracticeAreaTeamMember"
-VERIFIED_USER = "VerifiedUser"
-
